+++ front-end/config/_base.js
@@ -5,7 +5,7 @@ import { argv } from 'yargs'
 
 const debug = _debug('app:config:_base')
 const config = {
-  env : process.env.NODE_ENV = 'production',
+  env : process.env.NODE_ENV || 'development',
 
   // ----------------------------------
   // Project Structure
@@ -19,10 +19,10 @@ const config = {
   // ----------------------------------
   // Server Configuration
   // ----------------------------------
- // server_host: 'localhost',
- // server_port: process.env.PORT || 5000,
-  server_host : 's6ie1702.gel.usherbrooke.ca',
-  server_port : process.env.PORT || 80,
+  server_host: process.env.HOST || 'localhost',
+  server_port: process.env.PORT || 5000,
+ // server_host : 's6ie1702.gel.usherbrooke.ca',
+ // server_port : process.env.PORT || 80,
 
   // ----------------------------------
   // Compiler Configuration
