+++ front-end/src/components/header.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 
-import logo from '../../images/logo.svg'
+//import logo from '../../images/logo.svg'
 
 
 export default class InvoicePreviewHeader extends Component {
