+++ front-end/src/styles/core.scss
@@ -18,7 +18,7 @@
 // Modules
 
 @import
-"modules/_form",
+//"modules/_form",
 "modules/_icon",
 "modules/_item",
 "modules/_button";
