+++ front-end/src/views/HomeView/HomeView.js
@@ -1,13 +1,17 @@
 import React, { Component } from 'react'
 
 
-import FileList from '../../components/file-list'
+//import Depots from '../../components/depot-list'
+import Files from '../../components/file-list'
+
+import Header from '../../components/header'
 
 export default class HomeView extends Component {
   render () {
     return (
       <div className='container'>
-        <FileList/>
+        <Header />
+        <Files/>
       </div>
     )
   }
